Finished with FileParser
Done with File Parser class. Used a lexer to tokenize the lines in file. Still needs some more testing (e.g. make sure it takes .txt files as input and look for edge cases that makes the parser think the file is correctly structured when it's not), but it passed enough tests of mine that I think it's good enough to push.

What I need to work on next is the explorer.py file, which is setting up
the logic for finding all files in a directory and manipulating them.