Finished with explorer and perhaps the project
Took me about 2 days to finish this project. The code is pretty easy to
follow, especially with all the comments I added.

I removed the stats.py and file_creator.py files because I decided that we can add their functionality into the explorer.py file. The explorer file wasn't going to be too large, so I added the features that were going to be in stats and file_creator into explorer instead and then deleted them.

I added to the README file, though I used ChatGPT to help me out with it
because I was too tired to write it.

I don't know if I will be committing anymore to this repo, but if
someone recommends an changes or finds bugs, I'll make time for it.